+++ database/migrations/2015_12_10_144603_create_question_choices_table.php
@@ -14,8 +14,8 @@ class CreateQuestionChoicesTable extends Migration
     public function up()
     {
         Schema::create('question_choices', function (Blueprint $table) {
-            $table->bigIncrements('id');
-            $table->unsignedBigInteger('question_id');
+            $table->increments('id');
+            $table->unsignedInteger('question_id');
             $table->foreign('question_id')
                 ->references('id')->on('questions')
                 ->onDelete('cascade');
