+++ database/migrations/2015_12_10_144727_create_questions_badges_table.php
@@ -14,11 +14,11 @@ class CreateQuestionsBadgesTable extends Migration
     public function up()
     {
         Schema::create('questions_badges', function (Blueprint $table) {
-            $table->unsignedBigInteger('question_id');
+            $table->unsignedInteger('question_id');
             $table->foreign('question_id')
                 ->references('id')->on('questions')
                 ->onDelete('cascade');
-            $table->unsignedBigInteger('badge_id');
+            $table->unsignedInteger('badge_id');
             $table->foreign('badge_id')
                 ->references('id')->on('badges')
                 ->onDelete('cascade');
